[GP-94] Update logic for rendering
