+++ src/features/reading/api.js
@@ -7,7 +7,7 @@ const client = axios.create({
   },
 });
 
-export const fetchReadingQuestions = async () => {
+export const getReadingQuestions = async () => {
   const response = await client.get(
     "/topics/ef6b69aa-2ec2-4c65-bf48-294fd12e13fc?skillName=READING"
   );
